First CYA page now a hub
Added in all the logic so the user can change an answer from the CYA page, and as long as the change doesn't affect anything else, the Continue button will take them back to the CYA page.

Also added in all the logic for the Back links.

Noticing we don't have a back link on the task list page. Wondering if this affected usability testing.